Added shade documentation and reformatted to match example backends
