+++ bin/make_manual.py
@@ -83,6 +83,7 @@ docs = [
     "protondrive.md",
     "seafile.md",
     "sftp.md",
+    "shade.md"
     "smb.md",
     "storj.md",
     "sugarsync.md",
