+++ docs/content/local.md
@@ -641,7 +641,7 @@ Properties:
 - Config:      encoding
 - Env Var:     RCLONE_LOCAL_ENCODING
 - Type:        Encoding
-- Default:     Slash,Dot
+- Default:     Slash,InvalidUtf8,Dot
 
 #### --local-description
